% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bonous.R
\docType{package}
\name{BONUS}
\alias{BONUS}
\alias{BONUS-package}
\title{A package for modeling ridge regression}
\description{
Ridge regression can be a good alternative when we have a lot of covariates (when p > n) or in the
situation of multicollinearity. More information on ridge regression can be found in chapter 3.4.1 in [1].
}
\author{
Nour Qweder until nu ^_^
}
