% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeregClass.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Ridge regression class}
\description{
ridgereg(formula, data, lambda).
Ridge regression can be a good alternative when we have a lot of covariates (when p > n) or in the
situation of multicollinearity. More information on ridge regression can be found in chapter 3.4.1 in [1].
}
\section{Fields}{

\describe{
\item{\code{lambda}}{argument lambda a constant for ridge}

\item{\code{myFormula}}{formula, is a object of class formula.}

\item{\code{myData}}{character, The data is typically a data.frame. The gieven data set.}
}}


\references{
Reference Classes: \url{http://adv-r.had.co.nz/R5.html}

QR decomposition: \url{https://en.wikipedia.org/wiki/QR_decomposition}

Vignettes: \url{http://r-pkgs.had.co.nz/vignettes.html}
}
